+++ src/MrKWatkins.OakCpu.Z80.TestSuites/InstructionLevel/TestEvent.cs
@@ -1,12 +1,12 @@
 namespace MrKWatkins.OakCpu.Z80.TestSuites.InstructionLevel;
 
 [SuppressMessage("ReSharper", "InconsistentNaming")]
-public sealed record TestEvent(TestEventType Type, int TState, ushort Address, byte Data)
+public sealed record TestEvent(TestEventType Type, ulong TState, ushort Address, byte Data)
 {
     /// <summary>
     /// The T-state which generated the event.
     /// </summary>
-    public int TState { get; } = TState;
+    public ulong TState { get; } = TState;
 
     public override string ToString() => $"{TState}: {Type}, 0x{Address:X4} 0x{Data:X2}";
 }
